facenet/code/requirements.txt
facenet-pytorch
torch
torchvision
pillow
